# Conflicts:
#	Content/ProjectG/Character/BP_PlayerCharacter.uasset
#	Content/ProjectG/Levels/TestLevel.umap
#	Source/ProjectG/Character/PGPlayerCharacter.cpp
#	Source/ProjectG/Character/PGPlayerCharacter.h

+++ Source/ProjectG/Item/ItemActor/PGProjectileItemBase.cpp
@@ -35,6 +35,6 @@ void APGProjectileItemBase::OnHit(UPrimitiveComponent* HitComponent,
 {
 }
 
-void APGProjectileItemBase::PlaySound_Implementation(const FVector& HitLocation, float HitPower)
+void APGProjectileItemBase::PlaySound_Implementation(const FVector& HitLocation, uint8 HitPower)
 {
 }
